*Stub.php moved
*Stub.php files moved.
Fixed namespaces
Tests OK
