+++ tests/OAuth2/Tests/Fixtures/OAuth2GrantUserStub.php
@@ -1,13 +1,17 @@
 <?php
 
-namespace OAuth2;
+namespace OAuth2\Tests\Fixtures;
 
 use OAuth2\Model\IOAuth2Client;
+use OAuth2\IOAuth2GrantUser;
+use OAuth2\Tests\Fixtures\OAuth2StorageStub;
 
-class OAuth2GrantUserStub extends OAuth2StorageStub implements IOAuth2GrantUser {
+class OAuth2GrantUserStub extends OAuth2StorageStub implements IOAuth2GrantUser
+{
     private $users;
 
-    public function addUser($username, $password, $scope = null, $data = null) {
+    public function addUser($username, $password, $scope = null, $data = null)
+    {
         $this->users[$username] = array(
             'password' => $password,
             'scope' => $scope,
@@ -15,7 +19,8 @@ class OAuth2GrantUserStub extends OAuth2StorageStub implements IOAuth2GrantUser 
         );
     }
 
-    public function checkUserCredentials(IOAuth2Client $client, $username, $password) {
+    public function checkUserCredentials(IOAuth2Client $client, $username, $password)
+    {
         if (!isset($this->users[$username])) {
             return false;
         }
@@ -25,7 +30,7 @@ class OAuth2GrantUserStub extends OAuth2StorageStub implements IOAuth2GrantUser 
                 'data' => $this->users[$username]['data'],
             );
         }
+
         return false;
     }
 }
-
